refactor(access-service): extract observe/write surface into access-observe-write-surface.ts
God-file decomposition, fourth access-service cluster. Moves the
write-side and operational surfaces verbatim (879 LOC, 10 methods):
runObserve (turn ingestion incl. idempotency + scope resolution),
memoryStore, suggestionSubmit, memoryActionApply, workTask/workProject/
workBoard, briefing, qmdHealth, namespaceQmdHealth into
AccessObserveWriteSurface (14 deps via selfDeps live wiring).

access-service.ts: 6,796 -> 6,050 LOC (9,038 at program start).
